Update env.yml to match website for students.
In particular, remove unix-only explicit dependencies that were causing trouble
on Windows. They will be OK on linux, as they get pulled implicitly by other
packages, but we don't want to list them as hard deps for the environment.

Also, reinstate the sqlalchemy dep that got accidentally deleted as part of
reverting a bigger PR.